refactor/api_data: UserRow component
