+++ client/src/pages/administrator/user/ManageUser.tsx
@@ -1,13 +1,5 @@
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import {
-  useGetUsersQuery,
-  useSoftDeleteUserMutation,
-  GetUsersDocument,
-  useRestoreUserMutation,
-} from "../../../types/graphql-types";
-import useNotification from "../../../hooks/useNotification";
-import BtnCrud from "../../../components/BtnCrud";
+import { useGetUsersQuery } from "../../../types/graphql-types";
 import BtnLink from "../../../components/BtnLink";
 import PageTitle from "../../../components/PageTitle";
 import Table from "@mui/material/Table";
@@ -19,16 +11,12 @@ import TableRow from "@mui/material/TableRow";
 import Paper from "@mui/material/Paper";
 import Stack from "@mui/material/Stack";
 import Pagination from "@mui/material/Pagination";
+import UserRow from "../../../components/user/UserRow";
 
 export default function ManageUser() {
   const [page, setPage] = useState<number>(1);
   const [limit] = useState<number>(10);
   const offset = (page - 1) * limit;
-
-  const navigate = useNavigate();
-
-  // User feedback
-  const { notifySuccess, notifyError } = useNotification();
 
   const { loading, error, data } = useGetUsersQuery({
     variables: {
@@ -37,67 +25,12 @@ export default function ManageUser() {
     },
   });
 
-  const [softDeleteUserMutation] = useSoftDeleteUserMutation();
-  const [restoreUserMutation] = useRestoreUserMutation();
-
   // the event parameter is prefixed with _ to avoid the linter flag as event is not used here but is mandatory in the MUI Pagination component
   const handlePageChange = (
     _event: React.ChangeEvent<unknown>,
     value: number,
   ) => {
     setPage(value);
-  };
-
-  const handleSoftDeleteUser = async (userId: number) => {
-    try {
-      await softDeleteUserMutation({
-        variables: {
-          data: { id: userId },
-        },
-        refetchQueries: [
-          {
-            query: GetUsersDocument,
-            variables: {
-              limit: limit,
-              offset: offset,
-            },
-          },
-        ],
-      });
-
-      notifySuccess("Utilisateur désactivé avec succès");
-    } catch (error) {
-      notifyError("Erreur lors de la désactivation de l'utilisateur");
-      console.error("Erreur lors de la désactivation de l'utilisateur", error);
-    }
-  };
-
-  const handleRestoreUser = async (userId: number) => {
-    try {
-      await restoreUserMutation({
-        variables: {
-          data: { id: userId },
-        },
-        refetchQueries: [
-          {
-            query: GetUsersDocument,
-            variables: {
-              limit: limit,
-              offset: offset,
-            },
-          },
-        ],
-      });
-
-      notifySuccess("Utilisateur réactivé avec succès");
-    } catch (error) {
-      notifyError("Erreur lors de la réactivé de l'utilisateur");
-      console.error("Erreur lors de la réactivé de l'utilisateur", error);
-    }
-  };
-
-  const handleEditUser = (userId: number) => {
-    navigate(`/administrator/user/edit/${userId}`);
   };
 
   if (loading) return <p>🥁 Chargement...</p>;
@@ -142,43 +75,12 @@ export default function ManageUser() {
           <TableBody>
             {data &&
               data.getUsers.users.map((user) => (
-                <TableRow
+                <UserRow
                   key={user.id}
-                  sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
-                >
-                  <TableCell component="th" scope="row">
-                    {user.id}
-                  </TableCell>
-                  <TableCell align="left">{user.firstname}</TableCell>
-                  <TableCell align="left">{user.lastname}</TableCell>
-                  <TableCell align="left">
-                    {user.email}
-                    <br />
-                    {user.roles.map((role) => role.label)}
-                  </TableCell>
-                  <TableCell align="right">
-                    <Stack spacing={2} direction="row">
-                      <BtnCrud
-                        disabled={false}
-                        handleClick={() => handleEditUser(user.id)}
-                        type={"edit"}
-                      />
-                      {user.deletedAt === null ? (
-                        <BtnCrud
-                          disabled={false}
-                          handleClick={() => handleSoftDeleteUser(user.id)}
-                          type={"disable"}
-                        />
-                      ) : (
-                        <BtnCrud
-                          disabled={false}
-                          handleClick={() => handleRestoreUser(user.id)}
-                          type={"enable"}
-                        />
-                      )}
-                    </Stack>
-                  </TableCell>
-                </TableRow>
+                  user={user}
+                  offset={offset}
+                  limit={limit}
+                />
               ))}
           </TableBody>
         </Table>
